Created define_constraint()
Defines a constraint and saves its information in the corresponding
lists. This function does not add the constraints to the cplex model.
The reason I'm using these lists is because addign variables and
constraints in batches is faster than adding them one by one for some
reason.

Args:
    constr_name (str): Name of the constraint.
    constr_linear_expr (cplex.SparsePair): Linear expression of the constraint.
    constr_sense (str): Sense of the constraint.
    constr_rhs (float): Right hand side of the constraint.

Plus documentations fixes